Route: Require exact matches when parsing
Instead of relying on route ordering, a parse should only be successful
if no path elements are left. This reverts to the behaviour pre-v0.2.0.

URLs will still parse if they contain parameters or a fragment that were
not specified in the route. However, this also can be restricted with
the newly introduced function `parseArgsStrict()`.

In cases where remaining path elements or parameters are expected, the
matchers `Elems` and `Params` were added.

Further changes include renaming of two `Route` functions:
1. `parse()` -> `parseArgs()`
2. `parseInternal()` -> `parse()`

Since these are potentially breaking changes and may require developers
to update their routing tables, the version number was increased to
v0.3.0.

Closes #37.